+++ spec/unit/hanami/utils/callbacks_spec.rb
@@ -279,7 +279,18 @@ RSpec.describe Hanami::Utils::Callbacks::Chain do
     end
 
     it "raises an error if try to add a callback when frozen" do
-      expect {  @chain.append :authenticate! }.to raise_error RuntimeError
+      expect { @chain.append :authenticate! }.to raise_error RuntimeError
+    end
+  end
+
+  describe "#==" do
+    it "is equal to another chain based on its contents" do
+      @chain.append :callback_a, :callback_b
+
+      other = @chain.class.new
+      other.append :callback_a, :callback_b
+
+      expect(@chain).to eq other
     end
   end
 end